+++ lib/src/deriv_chart/interactive_layer/crosshair/core/crosshair_details.dart
