+++ lib/src/deriv_chart/interactive_layer/crosshair/painters/crosshair_dot_painter.dart
