+++ lib/src/deriv_chart/interactive_layer/crosshair/painters/highlight/crosshair_highlight_painter.dart
