+++ lib/src/deriv_chart/interactive_layer/crosshair/painters/highlight/crosshair_hollow_candle_highlight_painter.dart
@@ -1,4 +1,4 @@
-import 'package:deriv_chart/src/deriv_chart/interactive_layer/crosshair/crosshair_highlight_painter.dart';
+import 'package:deriv_chart/src/deriv_chart/interactive_layer/crosshair/painters/highlight/crosshair_highlight_painter.dart';
 import 'package:deriv_chart/src/models/candle.dart';
 import 'package:flutter/material.dart';
