+++ lib/src/deriv_chart/interactive_layer/crosshair/painters/line/large_screen_crosshair_line_painter.dart
@@ -1,5 +1,5 @@
 import 'package:deriv_chart/src/deriv_chart/chart/helpers/paint_functions/paint_line.dart';
-import 'package:deriv_chart/src/deriv_chart/interactive_layer/crosshair/crosshair_line_painter.dart';
+import 'package:deriv_chart/src/deriv_chart/interactive_layer/crosshair/painters/line/crosshair_line_painter.dart';
 import 'package:flutter/material.dart';
 
 /// A custom painter to paint the crosshair `line` for large screens.
