+++ lib/src/deriv_chart/interactive_layer/crosshair/utils/find.dart
